@@ -5,6 +5,8 @@
   "author": "yaminncco <yaminncco@gmail.com>",
   "license": "MIT",
   "main": "dist/vue-sidebar-menu.js",
+  "module": "dist/vue-sidebar-menu.esm.js",
+  "unpkg": "dist/vue-sidebar-menu.min.js",
   "types": "./types/index.d.ts",
   "files": [
     "dist",
@@ -26,14 +28,14 @@
     "admin-sidebar"
   ],
   "scripts": {
-    "dev": "cross-env NODE_ENV=development webpack-dev-server --hot --open --config webpack.config.dev.js",
-    "build:demo": "cross-env NODE_ENV=production webpack --progress --hide-modules --config webpack.config.dev.js",
-    "build:lib": "cross-env NODE_ENV=production webpack --progress --hide-modules --config webpack.config.prod.js",
+    "dev": "vite",
+    "build:demo": "vite build",
+    "build:lib": "rollup -c rollup.config.js",
     "lint": "eslint --ext .js,.vue ."
   },
   "dependencies": {
     "path-to-regexp": "^1.7.0",
-    "vue": "^2.5.21"
+    "vue": "^3.0.0"
   },
   "postcss": {
     "plugins": {
@@ -46,31 +48,26 @@
     "not ie <= 8"
   ],
   "devDependencies": {
-    "autoprefixer": "^9.4.4",
-    "babel-core": "^6.26.0",
-    "babel-loader": "^7.1.2",
-    "babel-preset-env": "^1.6.0",
-    "babel-preset-stage-3": "^6.24.1",
-    "cross-env": "^5.0.5",
-    "css-loader": "^2.1.1",
-    "eslint": "^5.14.1",
+    "@babel/core": "^7.11.6",
+    "@babel/preset-env": "^7.11.5",
+    "@rollup/plugin-babel": "^5.2.1",
+    "@rollup/plugin-node-resolve": "^9.0.0",
+    "@vue/compiler-sfc": "^3.0.0",
+    "autoprefixer": "^9.8.6",
+    "eslint": "^5.16.0",
     "eslint-config-standard": "^12.0.0",
-    "eslint-plugin-import": "^2.16.0",
+    "eslint-plugin-import": "^2.22.0",
     "eslint-plugin-node": "^8.0.1",
-    "eslint-plugin-promise": "^4.0.1",
-    "eslint-plugin-standard": "^4.0.0",
-    "eslint-plugin-vue": "^5.2.2",
-    "file-loader": "^1.1.4",
-    "mini-css-extract-plugin": "^0.5.0",
-    "node-sass": "^4.14.1",
-    "postcss-loader": "^3.0.0",
-    "sass-loader": "^6.0.6",
-    "vue-loader": "^15.5.0",
-    "vue-router": "^3.0.2",
-    "vue-template-compiler": "^2.5.21",
-    "webpack": "^4.42.0",
-    "webpack-cli": "^3.3.11",
-    "webpack-dev-server": "^3.11.0",
-    "webpack-merge": "^4.2.1"
+    "eslint-plugin-promise": "^4.2.1",
+    "eslint-plugin-standard": "^4.0.1",
+    "eslint-plugin-vue": "^5.2.3",
+    "rollup": "^2.28.2",
+    "rollup-plugin-commonjs": "^10.1.0",
+    "rollup-plugin-postcss": "^3.1.8",
+    "rollup-plugin-terser": "^7.0.2",
+    "rollup-plugin-vue": "^6.0.0-beta.10",
+    "sass": "^1.26.11",
+    "vite": "^1.0.0-rc.4",
+    "vue-router": "^4.0.0-beta.12"
   }
 }
