@@ -1,7 +1,7 @@
-import Vue from 'vue'
+import { createApp } from 'vue'
 import App from './App.vue'
-import VueRouter from 'vue-router'
-import VueSidebarMenu from '../src/index'
+import { createRouter, createWebHashHistory } from 'vue-router'
+import VueSidebarMenu from '/@vue-sidebar-menu/'
 
 import Installation from './components/Installation.vue'
 import BasicUsage from './components/BasicUsage.vue'
@@ -9,10 +9,8 @@ import Props from './components/Props.vue'
 import Events from './components/Events.vue'
 import Styling from './components/Styling.vue'
 
-Vue.use(VueRouter)
-Vue.use(VueSidebarMenu)
-
-const router = new VueRouter({
+const router = createRouter({
+  history: createWebHashHistory(),
   routes: [
     {
       path: '/',
@@ -42,8 +40,7 @@ const router = new VueRouter({
   ]
 })
 
-new Vue({ // eslint-disable-line no-new
-  el: '#app',
-  router,
-  render: h => h(App)
-})
+createApp(App)
+  .use(router)
+  .use(VueSidebarMenu)
+  .mount('#app')
